Fix global variable declarations/definitions for GCC 10 compatibility
This boils down to declaring global variables as "extern" in header files
and defining them in just one place.

See https://www.gnu.org/software/gcc/gcc-10/porting_to.html
